requirements.txt
fastapi
uvicorn
python-dotenv
pymongo[srv]
cloudinary
python-multipart

